pyproject.toml
[project]
name = "chem_datasets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.85",
    "dask>=2025.3.0",
    "datasets>=3.5.0",
    "mpi4py>=4.0.3",
    "numpy>=2.2.4",
    "openbabel-wheel>=3.1.1.21",
    "pyarrow>=19.0.1",
    "rdkit>=2024.9.6",
]
